move crypto/storage packages to pipeline pkg
and also add godocs pipeline pkg documentation
