+++ client/pipeline/crypto/doc.go
