+++ client/pipeline/crypto/hash.go
