+++ client/pipeline/crypto/hash_test.go
