+++ client/pipeline/crypto/sha.go
