+++ client/pipeline/crypto/sha_test.go
