+++ client/pipeline/pipeline.go
@@ -3,8 +3,8 @@ package pipeline
 import (
 	"io"
 
-	"github.com/zero-os/0-stor/client/components/crypto"
 	"github.com/zero-os/0-stor/client/metastor"
+	"github.com/zero-os/0-stor/client/pipeline/crypto"
 	"github.com/zero-os/0-stor/client/pipeline/processing"
 )
