+++ client/pipeline/storage/random.go
