+++ client/pipeline/storage/replicated.go
